Make CI fully green: mypy strict + typed package
- Add src/regime/py.typed marker so mypy sees our installed package
  as type-aware rather than treating every cross-module import as
  untyped (which produced 'Skipping analyzing regime.X' on every file)
- Fix mypy module-path ambiguity: switch from files=['src'] to
  mypy_path='src' + files=['src/regime'] so the package is found via
  exactly one path (was producing 'Source file found twice under
  different module names')
- Add disable_error_code=['import-untyped'] so third-party libs
  without py.typed don't fail strict mode
- Extend mypy overrides with polars, sklearn, prefect, streamlit,
  structlog, tenacity, typer, rich, fastapi, uvicorn, pyarrow
- Cast numpy return through np.asarray() in costs.py to satisfy
  'no-any-return' in strict mode
- str()-cast classification_report output in lightgbm_regime.py
- float()-cast np.log result in hmm_regime._bic
- Guard 'd is not None' before accessing .date() in yfinance _norm_row
- Use HistoricalRegimeStats import directly in implications/service.py
  instead of stats_mod attribute (which mypy couldn't resolve through
  the re-export indirection)
- Switch dict[str, object] to dict[str, Any] in api/sample.py so
  the FastAPI route handlers don't fight with mypy on Pydantic kwargs
- Strip unused 'type: ignore' comments from api/main.py and
  implications/service.py

+++ src/regime/models/hmm_regime.py
@@ -36,7 +36,7 @@ def _bic(model: GaussianHMM, x: np.ndarray) -> float:
         + model.n_components * n_features  # diag covs
     )
     ll = float(model.score(x))
-    return -2.0 * ll + n_params * np.log(n_obs)
+    return float(-2.0 * ll + n_params * np.log(n_obs))
 
 
 def fit_hmm(
